requirements.txt
strawberry-graphql-django
django
tqdm